moved RevaturePro tests into a new folder
